+++ minitools/researchers/__init__.py
@@ -2,6 +2,7 @@
 Researchers module for external data research.
 """
 
+from minitools.researchers.hf_papers import HFPaperStats, HFPapersResearcher
 from minitools.researchers.trend import TrendResearcher
 
-__all__ = ["TrendResearcher"]
+__all__ = ["TrendResearcher", "HFPapersResearcher", "HFPaperStats"]